Fix: modify code to make durable from exception - add Illegal excapetion in OrderAccessRole - modify equal in strategy object - move errorcode not to call application code in domain
