+++ src/main/java/com/fhsh/daitda/order/domain/exception/OrderErrorCode.java
@@ -1,4 +1,4 @@
-package com.fhsh.daitda.order.application.service;
+package com.fhsh.daitda.order.domain.exception;
 
 import org.springframework.http.HttpStatus;
 
@@ -10,7 +10,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum OrderErrorCode implements ErrorCode {
-	//external error
+	//external exception
 	DELIVERY_SERVICE_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "DELIVERY_SERVICE_ERROR"),
 
 	//NOT FOUND
